Move buildStreamSource to indexUtil
